chore: Show correctly the clients in the Add Itinerary interface
I have created the ClientArrayAdapter class.
